consumer-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
confluent-kafka
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests